refactor: :recycle: move markets stats inside a specific hook
