+++ src/views/ExchangeBillboards.tsx
@@ -1,14 +1,8 @@
-import { useMemo } from 'react';
-
-import { getChainRevenue } from '@/services';
-import { ResolutionString } from 'public/tradingview/charting_library';
-import { useQueries, useQuery } from 'react-query';
-import { shallowEqual, useSelector } from 'react-redux';
 import styled, { type AnyStyledComponent } from 'styled-components';
 
 import { STRING_KEYS } from '@/constants/localization';
 
-import { useDydxClient, useStringGetter } from '@/hooks';
+import { usePerpetualMarketsStats, useStringGetter } from '@/hooks';
 
 import { breakpoints } from '@/styles';
 import { layoutMixins } from '@/styles/layoutMixins';
@@ -16,10 +10,6 @@ import { layoutMixins } from '@/styles/layoutMixins';
 import { Output, OutputType } from '@/components/Output';
 import { Tag } from '@/components/Tag';
 import { SparklineChart } from '@/components/visx/SparklineChart';
-
-import { getPerpetualMarkets } from '@/state/perpetualsSelectors';
-
-import { log } from '@/lib/telemetry';
 
 type ExchangeBillboardsProps = {
   className?: string;
@@ -32,108 +22,12 @@ startDate.setDate(startDate.getDate() - 1);
 export const ExchangeBillboards: React.FC<ExchangeBillboardsProps> = ({ className }) => {
   const stringGetter = useStringGetter();
 
-  const perpetualMarkets = useSelector(getPerpetualMarkets, shallowEqual) ?? {};
-  const { getCandles, compositeClient } = useDydxClient();
-
-  const markets = useMemo(
-    () => Object.values(perpetualMarkets).filter(Boolean),
-    [perpetualMarkets]
-  );
-
-  const { data } = useQuery({
-    queryKey: ['chain-revenue', startDate.toISOString(), endDate.toISOString()],
-    queryFn: () => {
-      try {
-        return getChainRevenue({
-          startDate,
-          endDate,
-        });
-      } catch (error) {
-        log('ExchangeBillboards getChainRevenue', error);
-      }
-    },
-    refetchOnWindowFocus: false,
-  });
-
-  const feeEarned = useMemo(() => data?.[0].total, [data]);
-  const feeEarnedChart = useMemo(
-    () =>
-      data?.map((point, x) => ({
-        x: x + 1,
-        y: point.total,
-      })) ?? [],
-    [data]
-  );
-
-  const results = useQueries(
-    markets.map((market) => ({
-      enabled: !!compositeClient && markets.length > 0,
-      queryKey: ['perpetualMarketCandles', market.id, '1HOUR'],
-      queryFn: () => {
-        try {
-          return getCandles({
-            marketId: market.id,
-            resolution: '60' as ResolutionString,
-            limit: 24,
-          });
-        } catch (error) {
-          log('ExchangeBillboards getCandles', error);
-        }
-      },
-      refetchOnWindowFocus: false,
-    }))
-  );
-
-  const volume24HUSDCChart = useMemo(() => {
-    const data = results.map((result) => result.data);
-
-    if (data && data[0]) {
-      const sum: number[] = data[0].map((_, columnIndex) =>
-        data.reduce((acc, row) => acc + parseFloat(row?.[columnIndex].usdVolume ?? '0'), 0)
-      );
-
-      const candles = sum.map((y, x) => ({ x: x + 1, y }));
-
-      return candles;
-    }
-
-    return [];
-  }, [results]);
-
-  const openInterestUSDCChart = useMemo(() => {
-    const data = results.map((result) => result.data);
-
-    if (data && data[0]) {
-      const sum: number[] = data[0].map((_, columnIndex) =>
-        data.reduce(
-          (acc, row) => acc + parseFloat(row?.[columnIndex].startingOpenInterest ?? '0'),
-          0
-        )
-      );
-
-      const candles = sum.map((y, x) => ({ x: x + 1, y }));
-
-      return candles;
-    }
-
-    return [];
-  }, [results]);
-
-  const { volume24HUSDC, openInterestUSDC } = useMemo(() => {
-    let volume24HUSDC = 0;
-    let openInterestUSDC = 0;
-
-    for (const { oraclePrice, perpetual } of markets) {
-      const { volume24H, openInterest = 0 } = perpetual || {};
-      volume24HUSDC += volume24H ?? 0;
-      if (oraclePrice) openInterestUSDC += openInterest * oraclePrice;
-    }
-
-    return {
-      volume24HUSDC,
-      openInterestUSDC,
-    };
-  }, [markets]);
+  const {
+    stats: { volume24HUSDC, openInterestUSDC, feeEarned },
+    volume24HUSDCChart,
+    openInterestUSDCChart,
+    feeEarnedChart,
+  } = usePerpetualMarketsStats();
 
   return (
     <Styled.MarketBillboardsWrapper className={className}>
